Crystal's point size is a cell height, and its dates defer to the machine
Two font-and-format fixes that together move this suite further than
everything before them put together.

The first: Crystal's font size is a character cell height and RDL's
FontSize is an em, and this converter emitted one as the other. The
resulting 11% glyph-width difference had been written down here as a
permanent ceiling -- a renderer disagreement neither side could fix,
because our text matches Arial's published metrics. It does. The em was
the wrong measurement to match.

Crystal's own PDF export settles it without inference: a 10pt Arial
object is written "8.95 Tf", 16pt as "14.3", 33pt as "29.55", every one
the nominal size times 0.8951 -- Arial's
unitsPerEm/(usWinAscent+usWinDescent). Crystal hands the point size to
GDI as a positive lfHeight, which GDI reads as cell height and then
picks whatever em makes the cell that tall. The ratio is the font's own,
not a constant: Arial 0.8951, Times New Roman 0.9030, Verdana 0.8228,
Calibri 0.8192, Arial Black 0.7091. From OS/2's winAscent/winDescent
because GDI's tmHeight is built from those; they match hhea for every
family here except Calibri, whose hhea would give 1.0. Checked against a
second family first: Crystal writes 6.6 and 8.25 for SalesByCustomer's
8pt and 10pt Verdana, predicted 6.58 and 8.23. A family not in the table
is emitted unchanged, since its ratio is its own font's.

Riding along: the font record carries the size twice, the whole-point
byte this parser read and the same size in twips four fields later. The
byte loses every half-point size -- 2,581 private records disagree with
it, among them 5.95pt (1,090), 7.95pt (614) and 10.5pt (258). That half
has no test and cannot have one here: every size in the public corpus is
a whole number of points, so byte and twips agree on all 3,142 of its
font records. Said so in the parser rather than shipping a test that
cannot fail.

The second: a date field whose record configures no date components
defers to the machine's short date, and was being given a fixed
MM/dd/yyyy. Ground truth came from Crystal's CSV export, which writes
rendered values -- exporting a report hands back its formatted dates as
strings, so bytes pair directly with behaviour. Seven public reports
display a date field; bytes[2] and bytes[3] of the format record are
what separate "04/24/2001" from "2000-12-03  12:00:00AM".

I nearly shipped this machine's locale as a format. Reading those ISO
strings as report-specified, I wrote yyyy-MM-dd literals and measured a
further 1.7 and 1.1 points for them. An existing test then failed,
asserting its report must carry no format because "order 1 means use
whatever short date the machine has". It was right: this box runs en-CA,
whose short date is yyyy-MM-dd. Every ISO date in that evidence was the
machine, not the report, and the visual suite would have rewarded baking
it in. The fix is to emit no format at all -- our engine and Crystal
both fall back to the machine for an unformatted date, so they agree by
deferring rather than by matching a constant.

At least thirteen date patterns exist in the private corpus and three
renderings cannot decide between them, so the rest is left alone. The
CSV route makes more ground truth cheap to collect.

Measured. Font: all 88 of 88 public reports and 2,316 of 2,324 private
emit different RDL. Dates: a further 5 public and 83 private.

  CustomerList             61.0 -> 90.2
  boyum__SampleReport      57.0 -> 80.5
  BeforeTV                 57.4 -> 75.2
  Orders10k                57.2 -> 74.4
  ProductPriceList         62.8 -> 63.3
  Country-Region-Sort      60.9 -> 62.2
  Orders5-150              51.1 -> 59.9
  SalesByCustomer-Grouped  54.8 -> 54.0   (inside tolerance)

0 fatal and 0 exceptions across both corpora, 100% parse. 923 crystal
tests, 5 RptEngine, 12 visual + 1 ignored. Teeth-checked: reverting the
converter fails 6, reverting the parser fails 2.

+++ src/Majorsilence.Crystal.Model/Objects/ObjectFormat.cs
@@ -3,7 +3,13 @@ namespace Majorsilence.Crystal.Model.Objects;
 public sealed class ObjectFormat
 {
     public string? FontName { get; init; }
-    public int? FontSize { get; init; }
+    /// <summary>
+    /// Font size in points, as Crystal means it: the height of the character cell, not the
+    /// em square. The two are not the same number, and the converter turns one into the
+    /// other - see RdlConverter.EmPointsFor. Fractional because the file records the size in
+    /// twips and half-point sizes are common.
+    /// </summary>
+    public double? FontSize { get; init; }
     public bool Bold { get; init; }
     public bool Italic { get; init; }
     public bool Underline { get; init; }
